requirements.txt
aiofiles==23.2.1
alembic==1.16.1
amqp==5.3.1
annotated-types==0.6.0
anyio==4.3.0
APScheduler==3.10.4
arrow==1.3.0
asn1==2.7.0
asttokens==2.4.1
bcrypt==4.0.1
beautifulsoup4==4.12.2
billiard==4.2.1
blessed==1.20.0
blinker==1.9.0
bpylist2==4.1.1
bs4==0.0.2
cached-property==1.5.2
celery==5.3.4
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coloredlogs==15.0.1
construct==2.10.69
cryptography==42.0.5
customtkinter==5.2.2
daemonize==2.5.0
darkdetect==0.8.0
decorator==5.1.1
developer-disk-image==0.0.2
dnspython==2.6.1
email-validator==2.1.0
enum-compat==0.0.3
executing==2.0.1
fastapi==0.110.0
Flask==2.3.3
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.0.5
gpxpy==1.5.0
gunicorn==21.2.0
h11==0.14.0
hexdump==3.3
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
humanfriendly==10.0
hyperframe==6.0.1
icalendar==6.3.0
idna==3.6
ifaddr==0.2.0
iniconfig==2.1.0
inquirer3==0.4.0
ipsw-parser==1.2.2
ipython==8.22.2
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
kombu==5.5.4
la-panic==0.4.9
Mako==1.3.10
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
nest-asyncio==1.6.0
netifaces==0.11.0
numpy==1.24.3
opack==0.1.0
opencv-python==4.8.1.78
orjson==3.10.0
packaging==24.0
parameter-decorators==0.0.2
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
pluggy==1.6.0
plumbum==1.8.2
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pycrashreport==1.2.2
pycryptodome==3.20.0
pydantic==2.6.4
pydantic-extra-types==2.6.0
pydantic-settings==2.2.1
pydantic_core==2.16.3
Pygments==2.17.2
pygnuutils==0.1.1
pyimg4==0.8
PyJWT==2.10.1
pykdebugparser==1.2.4
pymobiledevice3==2.30.0
pytest==8.4.0
pytest-mock==3.14.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-editor==1.0.4
python-multipart==0.0.9
pytun-pmd3==1.0.0
pytz==2025.2
pyusb==1.2.1
PyYAML==6.0.1
qh3==0.15.1
readchar==4.0.6
redis==5.0.1
remotezip==0.12.3
requests==2.31.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.23
srptools==1.0.1
sslpsk-pmd3==1.0.3
stack-data==0.6.3
starlette==0.36.3
termcolor==2.4.0
tk==0.1.0
tqdm==4.66.2
traitlets==5.14.2
ttkthemes==3.2.2
types-python-dateutil==2.9.0.20240316
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
ujson==5.9.0
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
vine==5.1.0
watchfiles==0.21.0
wcwidth==0.2.13
websockets==12.0
Werkzeug==2.3.7
wsproto==1.2.0
xonsh==0.15.1
zeroconf==0.131.0
